+++ tiled/utils.py
@@ -2,7 +2,6 @@ import base64
 import builtins
 import collections.abc
 import contextlib
-import enum
 import functools
 import importlib
 import importlib.util
@@ -302,11 +301,6 @@ class CachingMap(collections.abc.Mapping):
         )
 
 
-class SpecialUsers(str, enum.Enum):
-    public = "public"
-    admin = "admin"
-
-
 def _line(nodes, last):
     "Generate a single line for the tree utility"
     tee = "├"
@@ -429,6 +423,7 @@ class Sentinel:
 
 UNCHANGED = Sentinel("UNCHANGED")
 UNSET = Sentinel("UNSET")
+SingleUserPrincipal = Sentinel("SingleUserPrincipal")
 
 
 def import_object(colon_separated_string, accept_live_object=True):